+++ FlowMeter/FlowMeter.h
@@ -7,29 +7,29 @@
 
 class FlowMeter {
   public:
-    const static unsigned int MINIMUM_PULSES_BETWEEN_READ_TO_SET_FLOWING = 5;
+    const static uint8_t MINIMUM_PULSES_BETWEEN_READ_TO_SET_FLOWING = 5;
 
   private: 
-    const static unsigned int DEFAULT_FACTOR_PULSES_PER_LITER = 350;
-    const static unsigned int TICKER_LENGTH_MILISEC = 500;
-    const static unsigned int TRANSLATE_PULSES_TO_LITERS_AND_RESET = DEFAULT_FACTOR_PULSES_PER_LITER * 1000;
+    const static uint16_t DEFAULT_FACTOR_PULSES_PER_LITER = 350;
+    const static uint16_t TICKER_LENGTH_MILISEC = 500;
+    const static uint16_t TRANSLATE_PULSES_TO_LITERS_AND_RESET =  (uint16_t) (DEFAULT_FACTOR_PULSES_PER_LITER * 1000);
 
   public:
-    FlowMeter(int signalGpio, unsigned int calibrationFactor = DEFAULT_FACTOR_PULSES_PER_LITER);
+    FlowMeter(uint8_t signalGpio, uint16_t calibrationFactor = DEFAULT_FACTOR_PULSES_PER_LITER);
     ~FlowMeter() {}
 
     void start(void);
     void restart(void);
-    unsigned int getCountedLiters(bool restart = false);
+    uint16_t getCountedLiters(bool restart = false);
     void setIsFlowing(bool isFlowing);
     bool isFlowing();
     void run();
 
   protected:
-    int _signalPort;
-    unsigned int _pulsesPerLiter;
+    uint8_t _signalPort;
+    uint16_t _pulsesPerLiter;
     bool _isFluidFlowing;
-    unsigned int _litersCounted;
+    uint16_t _litersCounted;
 
   private:
     friend void countPulses(FlowMeter *);
